DEV-41 move presentation to infra
